Refactor: Move spinner component to separate file
