+++ pkg/tui/components.go
@@ -2,7 +2,6 @@ package tui
 
 import (
 	"strings"
-	"time"
 
 	"github.com/gdamore/tcell/v2"
 	"github.com/killallgit/ryan/pkg/chat"
@@ -247,83 +246,6 @@ func (sb StatusBar) WithModelViewData(totalModels int, totalSize int64) StatusBa
 	}
 }
 
-type SpinnerComponent struct {
-	IsVisible bool
-	Frame     int
-	StartTime time.Time
-	Text      string
-	Style     tcell.Style
-}
-
-// Using spinner character set #25 from github.com/briandowns/spinner
-var spinnerFrames = []string{"ｦ", "ｧ", "ｨ", "ｩ", "ｪ", "ｫ", "ｬ", "ｭ", "ｮ", "ｯ", "ｱ", "ｲ", "ｳ", "ｴ", "ｵ", "ｶ", "ｷ", "ｸ", "ｹ", "ｺ", "ｻ", "ｼ", "ｽ", "ｾ", "ｿ", "ﾀ", "ﾁ", "ﾂ", "ﾃ", "ﾄ", "ﾅ", "ﾆ", "ﾇ", "ﾈ", "ﾉ", "ﾊ", "ﾋ", "ﾌ", "ﾍ", "ﾎ", "ﾏ", "ﾐ", "ﾑ", "ﾒ", "ﾓ", "ﾔ", "ﾕ", "ﾖ", "ﾗ", "ﾘ", "ﾙ", "ﾚ", "ﾛ", "ﾜ", "ﾝ"}
-
-func NewSpinnerComponent() SpinnerComponent {
-	return SpinnerComponent{
-		IsVisible: false,
-		Frame:     0,
-		StartTime: time.Now(),
-		Text:      "",
-		Style:     tcell.StyleDefault.Foreground(tcell.ColorGray),
-	}
-}
-
-func (sc SpinnerComponent) WithVisibility(visible bool) SpinnerComponent {
-	spinner := SpinnerComponent{
-		IsVisible: visible,
-		Frame:     sc.Frame,
-		StartTime: sc.StartTime,
-		Text:      sc.Text,
-		Style:     sc.Style,
-	}
-
-	// Reset animation when becoming visible
-	if visible && !sc.IsVisible {
-		spinner.StartTime = time.Now()
-		spinner.Frame = 0
-	}
-
-	return spinner
-}
-
-func (sc SpinnerComponent) WithText(text string) SpinnerComponent {
-	return SpinnerComponent{
-		IsVisible: sc.IsVisible,
-		Frame:     sc.Frame,
-		StartTime: sc.StartTime,
-		Text:      text,
-		Style:     sc.Style,
-	}
-}
-
-func (sc SpinnerComponent) NextFrame() SpinnerComponent {
-	if !sc.IsVisible {
-		return sc
-	}
-
-	return SpinnerComponent{
-		IsVisible: sc.IsVisible,
-		Frame:     (sc.Frame + 1) % len(spinnerFrames),
-		StartTime: sc.StartTime,
-		Text:      sc.Text,
-		Style:     sc.Style,
-	}
-}
-
-func (sc SpinnerComponent) GetCurrentFrame() string {
-	if !sc.IsVisible {
-		return ""
-	}
-	return spinnerFrames[sc.Frame]
-}
-
-func (sc SpinnerComponent) GetDisplayText() string {
-	if !sc.IsVisible {
-		return ""
-	}
-	// Only return the spinner character, no text
-	return sc.GetCurrentFrame()
-}
 
 type AlertDisplay struct {
 	IsSpinnerVisible bool
@@ -390,7 +312,7 @@ func (ad AlertDisplay) NextSpinnerFrame() AlertDisplay {
 
 	return AlertDisplay{
 		IsSpinnerVisible: ad.IsSpinnerVisible,
-		SpinnerFrame:     (ad.SpinnerFrame + 1) % len(spinnerFrames),
+		SpinnerFrame:     (ad.SpinnerFrame + 1) % GetSpinnerFrameCount(),
 		SpinnerText:      ad.SpinnerText,
 		ErrorMessage:     ad.ErrorMessage,
 		Width:            ad.Width,
@@ -401,7 +323,7 @@ func (ad AlertDisplay) GetSpinnerFrame() string {
 	if !ad.IsSpinnerVisible {
 		return ""
 	}
-	return spinnerFrames[ad.SpinnerFrame]
+	return GetSpinnerFrame(ad.SpinnerFrame)
 }
 
 func (ad AlertDisplay) GetDisplayText() string {
